API Move rendering into its own class
